+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.2.2
+------------------------------
+*2019-12-10*
+
+### Fix
+- Types: Allow passing `window` when using `useEvent`
+
 ## 0.2.0
 ------------------------------
 *2019-11-03*